src/auth/requirements.txt
PyJWT>=2.8.0
python-dotenv>=1.0.0